+++ src/ProductSaleServices/ProductService.API/Controllers/ProductGroupController.cs
@@ -13,7 +13,7 @@ namespace ProductService.API.Controllers
 
         public ProductGroupController(ProductGroupManager pm) => this.pm = pm;
 
-        [Route("queries")]
+        [Route("queries/all")]
         [HttpGet]
         public async Task<IEnumerable<ProductGroup>> Get() => await pm.GetAllProductGroups();
 
@@ -21,12 +21,23 @@ namespace ProductService.API.Controllers
         [HttpGet]
         public async Task<IEnumerable<ProductGroup>> GetUnavailableProductGroups() => await pm.GetUnavailableProductGroups();
 
-        [Route("queries/ownerId/{ownerId}")]
+        [Route("queries/user")]
         [HttpGet]
-        public async Task<IEnumerable<ProductGroup>> GetProductGroupsByOwnerId(string ownerId) => await pm.GetProductGroupsByOwnerId(ownerId);
+        [ProducesResponseType(200)]
+        [ProducesResponseType(400)]
+        public async Task<ActionResult<IEnumerable<ProductGroup>>> GetProductGroupsByOwnerId() {
+            // Authentication result in X-Forwarded-User
+            Request.Headers.TryGetValue("X-Forwarded-User", out Microsoft.Extensions.Primitives.StringValues ownerId);
+
+            if (ownerId.ToString() == null)
+                return BadRequest();
+
+            var products = await pm.GetProductGroupsByOwnerId(ownerId);
+            return Ok(products);
+        }
 
 
-        [HttpGet("queries/id/{productGroupId}")]
+        [HttpGet("queries/{productGroupId}")]
         [ProducesResponseType(200)]
         [ProducesResponseType(404)]
         public async Task<ActionResult<Product>> Get(string productGroupId)
@@ -39,13 +50,27 @@ namespace ProductService.API.Controllers
 
         [HttpDelete("{productGroupId}")]
         [ProducesResponseType(204)]
+        [ProducesResponseType(400)]
+        [ProducesResponseType(401)]
         [ProducesResponseType(404)]
         public async Task<IActionResult> Delete(string productGroupId)
         {
+            // Authentication result in X-Forwarded-User
+            Request.Headers.TryGetValue("X-Forwarded-User", out Microsoft.Extensions.Primitives.StringValues ownerId);
+
+            if (ownerId.ToString() == null)
+                return BadRequest();
+
             var group = await pm.GetProductGroupOrNull(productGroupId);
+            
             if (group == null)
                 return NotFound();
+            
+            if (ownerId.ToString() != group.OwnerID)
+                return Unauthorized();
+
             var result = await pm.DeleteProductGroup(productGroupId);
+            
             if (!result)
                 return NotFound();
             return NoContent();
@@ -54,8 +79,18 @@ namespace ProductService.API.Controllers
         [HttpPost]
         [ProducesResponseType(201)]
         [ProducesResponseType(400)]
+        [ProducesResponseType(401)]
         public async Task<IActionResult> Create([FromBody] ProductGroup group)
         {
+            // Authentication result in X-Forwarded-User
+            Request.Headers.TryGetValue("X-Forwarded-User", out Microsoft.Extensions.Primitives.StringValues ownerId);
+
+            if (ownerId.ToString() == null)
+                return BadRequest();
+
+            if (group.OwnerID != ownerId.ToString())
+                return Unauthorized();
+
             group = await pm.CreateProductGroup(group);
             if (group.ID != string.Empty)
                 return CreatedAtAction(nameof(Get), new { id = group.ID }, group);
@@ -68,11 +103,19 @@ namespace ProductService.API.Controllers
         [ProducesResponseType(404)]
         public async Task<IActionResult> Update(string productGroupId, [FromBody] ProductGroup newProductGroup)
         {
-            if (productGroupId != newProductGroup.ID)
+            // Authentication result in X-Forwarded-User
+            Request.Headers.TryGetValue("X-Forwarded-User", out Microsoft.Extensions.Primitives.StringValues ownerId);
+
+            if (productGroupId != newProductGroup.ID || ownerId.ToString() == null)
                 return BadRequest();
+
             var productGroup = await pm.GetProductGroupOrNull(productGroupId);
+
             if (productGroup == null)
                 return NotFound();
+
+            if (productGroup.OwnerID != ownerId.ToString())
+                return Unauthorized();
 
             var result = await pm.UpdateProductGroup(productGroupId, productGroup, newProductGroup);
             if (!result)
